+++ philo/inc/args.h
@@ -17,11 +17,11 @@
 
 typedef struct s_main_args
 {
-	size_t	num_of_philos;
-	size_t	time_to_die;
-	size_t	time_to_eat;
-	size_t	time_to_sleep;
-	int		number_of_times_each_philosopher_must_eat;
+	int	num_of_philos;
+	int	time_to_die;
+	int	time_to_eat;
+	int	time_to_sleep;
+	int	number_of_times_each_philosopher_must_eat;
 }			t_main_args;
 
 #endif
