+++ philo/srcs/actors/sv/_on_shut_down.c
@@ -6,7 +6,7 @@
 /*   By: ttsubo <ttsubo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/17 21:50:17 by ttsubo            #+#    #+#             */
-/*   Updated: 2025/06/21 17:19:17 by ttsubo           ###   ########.fr       */
+/*   Updated: 2025/06/29 12:59:40 by ttsubo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	_on_shut_down(t_sv_actor *self)
 {
-	size_t		i;
+	int	i;
 
 	i = 0;
 	while (i < self->prop->args.num_of_philos)
